+++ +gamrywb/+app/launchVTResistanceApp.m
@@ -423,14 +423,7 @@ function varargout = launchVTResistanceApp(varargin)
     end
 
     function swapPlots()
-        topX = ddTopX.Value;
-        topY = ddTopY.Value;
-        botX = ddBotX.Value;
-        botY = ddBotY.Value;
-        ddTopX.Value = botX;
-        ddTopY.Value = botY;
-        ddBotX.Value = topX;
-        ddBotY.Value = topY;
+        gamrywb.ui.swapTopBottomPlotSelections(ddTopX, ddTopY, ddBotX, ddBotY);
         refreshPlots();
     end
 
@@ -440,15 +433,12 @@ function varargout = launchVTResistanceApp(varargin)
     end
 
     function restoreDefaultPlotSelections()
-        ddTopX.Value = 'Time (s)';
-        ddTopY.Value = 'VT: Vf vs time';
-        ddBotX.Value = 'Time (s)';
-        ddBotY.Value = 'IT: Im vs time';
+        gamrywb.ui.setTopBottomPlotSelections(ddTopX, ddTopY, ddBotX, ddBotY, ...
+            topPlotDefaults, bottomPlotDefaults);
     end
 
     function resetAxesToDefaultState()
-        gamrywb.ui.hardResetAxis(axTop, 'Top Plot');
-        gamrywb.ui.hardResetAxis(axBottom, 'Bottom Plot');
+        gamrywb.ui.resetTopBottomAxes(axTop, axBottom);
     end
 
     function exportResultsCSV()